Moves ExecuteDetailsTimings into its own file.
