+++ program-runtime/src/lib.rs
@@ -7,3 +7,4 @@ pub mod log_collector;
 pub mod native_loader;
 pub mod neon_evm_program;
 pub mod stable_log;
+pub mod timings;
